Seeders y ajustes en la carga masiva de pagos de las facturas

+++ database/migrations/2022_07_28_023929_create_payments_histories_table.php
@@ -18,7 +18,7 @@ class CreatePaymentsHistoriesTable extends Migration
             $table->foreignId("user_id")->constrained("users", "id");
             $table->foreignId("invoice_id")->constrained("invoices", "id");
             $table->date("date");
-            $table->integer("payment");
+            $table->double("payment", 8, 2);
             $table->timestamps();
         });
     }
